+++ adminSiteClient/AdminApp.tsx
@@ -35,8 +35,8 @@ import {
 import { LoadingBlocker, Modal } from "./Forms.js"
 import { AdminAppContext } from "./AdminAppContext.js"
 import { Base64 } from "js-base64"
-import { ExplorerCreatePage } from "../explorerAdminClient/ExplorerCreatePage.js"
-import { ExplorersIndexPage } from "../explorerAdminClient/ExplorersListPage.js"
+import { ExplorerCreatePage } from "./ExplorerCreatePage.js"
+import { ExplorersIndexPage } from "./ExplorersListPage.js"
 import { EXPLORERS_ROUTE_FOLDER } from "@ourworldindata/explorer"
 import { AdminLayout } from "./AdminLayout.js"
 import { BulkGrapherConfigEditorPage } from "./BulkGrapherConfigEditor.js"
