--- devTools/explorerTools/tsconfig.json
@@ -1,11 +1,0 @@
-{
-    "extends": "../tsconfigs/tsconfig.base.json",
-    "compilerOptions": {
-        "outDir": "../../itsJustJavascript/devTools/explorerTools",
-        "rootDir": "."
-    },
-    "references": [
-        { "path": "../../gitCms" },
-        { "path": "../../explorerAdminServer" }
-    ]
-}
